Move avatar worker out of legacy
